Removed some (but not all) of the line endings I accidentally added while in VS.
Fixed a bunch of issues related to compiling unittests on VS.

Having a problem linking against boost unit tests, so can't run the tests yet.

+++ include/cereal/types/deque.hpp
@@ -52,7 +52,7 @@ namespace cereal
     size_type size;
     ar( make_size_tag( size ) );
 
-    deque.resize( size );
+    deque.resize( static_cast<size_t>( size ) );
 
     for( auto & i : deque )
       ar( i );
